several changes

+++ db/migrate/20170127205430_change_product_price.rb
@@ -1,6 +1,6 @@
 class ChangeProductPrice < ActiveRecord::Migration
   def change
     # change_column :products, :price, :decimal, precision: 10, scale: 2
-    change_column :products, :price, :integer 
+    change_column :products, :price, :float 
   end
 end
